% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{abbrDeparse}
\alias{abbrDeparse}
\title{Abbreviated deparse function}
\usage{
abbrDeparse(x, width = 60)
}
\arguments{
\item{x}{Name to deparse, a character string.}

\item{width}{Field width, an integer.}
}
\value{
A character string.
}
\description{
Function used to abbreviate name of dataset in \code{\link{anova}} function.
Taken from \code{lme4}.
}
\references{
\insertRef{batesFittingLinearMixedEffects2015}{galamm}
}
\author{
Douglas M. Bates, Martin Maechler, Ben Bolker, and Steve Walker
}
\keyword{internal}
