% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{augment.smX}
\alias{augment.smX}
\title{Function for creating augmented model matrix.}
\usage{
augment.smX(sm, nobs, np)
}
\arguments{
\item{sm}{List of smooth objects, each of which returned from
\code{mgcv::smoothCon}.}

\item{nobs}{Number of observations.}

\item{np}{Number of penalized parameters.}
}
\value{
A scaled and augmented model matrix.
}
\description{
This is an internal function taken from \code{mgcv}.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon Wood
}
\keyword{internal}
