% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{clone.smooth.spec}
\alias{clone.smooth.spec}
\title{Internal function from mgcv}
\usage{
clone.smooth.spec(specb, spec)
}
\arguments{
\item{specb}{spec}

\item{spec}{spec}
}
\value{
value
}
\description{
Internal function from mgcv
}
\author{
Simon Wood
}
