% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Example dataset with three measurements of a single latent variable.}
\format{
An object of class \code{data.frame} with 300 rows and 4 columns.
}
\usage{
dat1
}
\description{
The underlying smooth curve for the latent trait has been generated
\code{mgcv::gamSim}.
}
\keyword{datasets}
