% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat2}
\alias{dat2}
\title{Example dataset with binomial response and two levels.}
\format{
A data frame with 1800 rows and 4 variables:
\describe{
  \item{id}{Subject ID}
  \item{y}{Number of successes}
  \item{trials}{Number of trials. Always equal to 1.}
  \item{x}{Explanatory variable}
}
}
\usage{
dat2
}
\description{
Simulated dataset similar to the one used in Section 3.3 of
\insertCite{rabe-hesketh2005;textual}{galamm}. This dataset can be used
to fit models with random slopes.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
