% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat3}
\alias{dat3}
\title{Example dataset with binomial response and two levels.}
\format{
A data frame with 1800 rows and 4 variables:
\describe{
  \item{id}{Subject ID}
  \item{y}{Number of successes}
  \item{trials}{Number of trials. Always equal to 1.}
  \item{x}{Explanatory variable}
}
}
\usage{
dat3
}
\description{
This dataset can be used to fit models with quadratic random slopes.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
