% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epilep}
\alias{epilep}
\title{Epilepsy Data}
\format{
\subsection{\code{epilep} A data frame with 236 rows and 7 columns:}{

\describe{
\item{subj}{Subject ID.}
\item{y}{Number of seizures.}
\item{treat}{Dummy variable for treatment group.}
\item{visit}{Time at visit.}
\item{v4}{Dummy for visit 4.}
\item{lage}{Logarithm of age.}
\item{lbas}{Logarithm of a quarter of the number of seizures in the eight
weeks preceding entry into the trial.}
}
}
}
\source{
\url{http://www.gllamm.org/books/readme.html#11.3}
}
\usage{
epilep
}
\description{
Longitudinal epilepsy data from
\insertCite{leppikControlledStudyProgabide1987;textual}{galamm}. This
documentation is based on Chapter 11.3 of
\insertCite{skrondalGeneralizedLatentVariable2004;textual}{galamm}, where
the dataset is used.
}
\references{
\insertAllCited{}
}
\seealso{
Other datasets:
\code{\link{cognition}},
\code{\link{diet}},
\code{\link{hsced}},
\code{\link{latent_covariates}},
\code{\link{latent_covariates_long}},
\code{\link{lifespan}},
\code{\link{mresp}},
\code{\link{mresp_hsced}}
}
\concept{datasets}
\keyword{datasets}
