% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.galamm}
\alias{fixef.galamm}
\alias{fixef}
\title{Extract Fixed Effects from galamm Object}
\usage{
\method{fixef}{galamm}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{
Fixed effects
}
\description{
Extract Fixed Effects from galamm Object
}
