% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  dispformula = NULL,
  weights = NULL,
  data,
  family = gaussian,
  family_mapping = NULL,
  load_var = NULL,
  load.var = NULL,
  lambda = NULL,
  factor = NULL,
  factor_interactions = NULL,
  na.action = getOption("na.action"),
  start = NULL,
  control = galamm_control()
)
}
\arguments{
\item{formula}{A formula specifying the model. Smooth terms are defined in
the style of the \code{mgcv} and \code{gamm4} packages, see
\insertCite{woodGeneralizedAdditiveModels2017}{galamm} for an
introduction. Random effects are specified using \code{lme4} syntax, which
is described in detail in
\insertCite{batesFittingLinearMixedEffects2015}{galamm}. Factor loadings
will also be part of the model formula, and is based on the syntax of the
\code{PLmixed} package
\insertCite{rockwoodEstimatingComplexMeasurement2019}{galamm}.}

\item{dispformula}{An optional formula object specifying an expression for the
residual variance. Defaults to \code{NULL}, corresponding to homoscedastic
errors. The formula is defined in \code{lme4} style; see vignettes and
examples for details.}

\item{weights}{Deprecated. Use \code{dispformula} instead.}

\item{data}{A data.frame containing all the variables specified by the model
formula, with the exception of factor loadings.}

\item{family}{A family object specifying the response distribution of the
model. Currently \code{gaussian}, \code{binomial}, and \code{poisson}
with canonical link functions are supported. Models with mixed response
types, in which responses are distributed according to one of the three
supported distributions, can be set up using the \code{\link[=gfam]{gfam()}} function.}

\item{family_mapping}{Deprecated. See the documentation to \code{\link[=gfam]{gfam()}} for how
to set up mixed response type models.}

\item{load_var}{Optional character specifying the name of the variable in
\code{data} identifying what the factors load onto. Default to \code{NULL},
which means that there are no loading variables. Argument is case
sensitive.}

\item{load.var}{Deprecated. Use \code{load_var} instead.}

\item{lambda}{Optional factor loading matrix. Numerical values indicate that
the given value is fixed, while \code{NA} means that the entry is a
parameter to be estimated. Numerical values can only take the values 0 or
\enumerate{
\item The number of columns of \code{lambda} must be identical to the number
of elements in \code{factor}. Defaults to \code{NULL}, which means that
there is no factor loading matrix. If \code{lambda} is provided as a
vector, it will be converted to a \code{matrix} with a single column.
}}

\item{factor}{Optional character vector whose \eqn{j}th entry corresponds to
the \eqn{j}th column of the corresponding matrix in \code{lambda}. The
number of elements in \code{factor} must be equal to the number of columns
in \code{lambda}. Defaults to \code{NULL}, which means that there are no
factor loadings. Argument is case sensitive.}

\item{factor_interactions}{Optional list of length equal to the number of
columns in \code{lambda}. Each list element should be a \code{formula}
object containing the write-hand side of a regression model, of the form
\code{~ x + z}. Defaults to \code{NULL}, which means that no factor
interactions are used.}

\item{na.action}{Character of length one specifying a function which
indicates what should happen when the data contains \code{NA}s. The
defaults is set to the \code{na.action} setting of \code{options}, which
can be seen with \code{options("na.action")}. The other alternatives are
\code{"na.fail"} or \code{"na.exclude"}, which means that the function
fails if there as \code{NA}s in \code{data}.}

\item{start}{Optional named list of starting values for parameters. Possible
names of list elements are \code{"theta"}, \code{"beta"}, \code{"lambda"},
and \code{"weights"}, all of should be numerical vectors with starting
values. Default to \code{NULL}, which means that some relatively sensible
defaults are used. Names of parameters must be given in all lower case.}

\item{control}{Optional control object for the optimization procedure of
class \code{galamm_control} resulting from calling
\code{\link{galamm_control}}. Defaults to \code{NULL}, which means that the
defaults of \code{\link{galamm_control}} are used.}
}
\value{
An object of type \code{galamm} as described in \link{galammObject}.
}
\description{
This function fits a generalized additive latent and mixed model
(GALAMMs), as described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm}.
The building blocks of these models are generalized additive mixed models
(GAMMs) \insertCite{woodGeneralizedAdditiveModels2017}{galamm}, of which
generalized linear mixed models
\insertCite{breslowApproximateInferenceGeneralized1993,harvilleMaximumLikelihoodApproaches1977,hendersonBestLinearUnbiased1975,lairdRandomEffectsModelsLongitudinal1982}{galamm}
are special cases. GALAMMs extend upon GAMMs by allowing factor structures,
as commonly used to model hypothesized latent traits underlying observed
measurements. In this sense, GALAMMs are an extension of generalized linear
latent and mixed models (GLLAMMs)
\insertCite{skrondalGeneralizedLatentVariable2004,rabe-heskethGeneralizedMultilevelStructural2004}{galamm}
which allows semiparametric estimation. The implemented algorithm used to
compute model estimates is described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm},
and is an extension of the algorithm used for fitting generalized linear
mixed models by the \code{lme4} package
\insertCite{batesFittingLinearMixedEffects2015}{galamm}. The syntax used to
define factor structures is based on that used by the \code{PLmixed}
package, which is detailed in
\insertCite{rockwoodEstimatingComplexMeasurement2019;textual}{galamm}.
}
\examples{
# Mixed response model ------------------------------------------------------

# The mresp dataset contains a mix of binomial and Gaussian responses.

# We need to estimate a factor loading which scales the two response types.
loading_matrix <- matrix(c(1, NA), ncol = 1)

# Define mapping to families.
families <- gfam(list(gaussian, binomial))

# Fit the model
mod <- galamm(
  formula = y ~ x + (0 + level | id),
  data = mresp,
  family = families,
  factor = "level",
  load_var = "itemgroup",
  lambda = loading_matrix
)

# Summary information
summary(mod)


# Heteroscedastic model -----------------------------------------------------
# Residuals allowed to differ according to the item variable
# We also set the initial value of the random intercept standard deviation
# to 1
mod <- galamm(
  formula = y ~ x + (1 | id), dispformula = ~ (1 | item),
  data = hsced, start = list(theta = 1)
)
summary(mod)

# Generalized additive mixed model with factor structures -------------------

# The cognition dataset contains simulated measurements of three latent
# time-dependent processes, corresponding to individuals' abilities in
# cognitive domains. We focus here on the first domain, and take a single
# random timepoint per person:
dat <- subset(cognition, domain == 1)
dat <- split(dat, f = dat$id)
dat <- lapply(dat, function(x) x[x$timepoint \%in\% sample(x$timepoint, 1), ])
dat <- do.call(rbind, dat)
dat$item <- factor(dat$item)

# At each timepoint there are three items measuring ability in the cognitive
# domain. We fix the factor loading for the first measurement to one, and
# estimate the remaining two. This is specified in the loading matrix.
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)

# We can now estimate the model.
mod <- galamm(
  formula = y ~ 0 + item + sl(x, factor = "loading") +
    (0 + loading | id),
  data = dat,
  load_var = "item",
  lambda = loading_matrix,
  factor = "loading"
)

# We can plot the estimated smooth term
plot_smooth(mod, shade = TRUE)


# Interaction between observed and latent covariates ------------------------
# Define the loading matrix
lambda <- matrix(c(1, NA, NA), ncol = 1)

# Define the regression functions, one for each row in the loading matrix
factor_interactions <- list(~1, ~1, ~x)

# Fit the model
mod <- galamm(
  formula = y ~ type + x:response + (0 + loading | id),
  data = latent_covariates,
  load_var = "type",
  lambda = lambda,
  factor = "loading",
  factor_interactions = factor_interactions
)

# The summary output now include an interaction between the latent variable
# and x, for predicting the third element in "type"
summary(mod)

}
\references{
\insertAllCited{}
}
\seealso{
Other modeling functions:
\code{\link{galammObject}},
\code{\link[=gfam]{gfam()}},
\code{\link{s}},
\code{\link{t2}}
}
\concept{modeling functions}
