% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{gam.setup}
\alias{gam.setup}
\title{Function for setting up GAM terms}
\usage{
gam.setup(
  formula,
  pterms,
  data = stop("No data supplied to gam.setup"),
  knots = NULL,
  select = FALSE,
  scale.penalty = TRUE,
  diagonal.penalty = FALSE,
  apply.by = TRUE,
  list.call = FALSE,
  modCon = 0
)
}
\arguments{
\item{formula}{Formula including smooth terms, but not \code{lme4} style
random effect terms.}

\item{pterms}{Parametric terms.}

\item{data}{Model data.}

\item{knots}{Knots specification.}

\item{select}{Passed on to the argument \code{null.space.penalty} of \code{mgcv::smoothCon}.}

\item{scale.penalty}{Boolean value passed on to \code{mgcv::smoothCon}.}

\item{diagonal.penalty}{not documented yet}

\item{apply.by}{not documented yet}

\item{list.call}{not documented yet}

\item{modCon}{not documented yet}
}
\value{
an object
}
\description{
This function is derived from \code{mgcv:::gam.setup}.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon N Wood
}
\keyword{internal}
