% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{gam.side}
\alias{gam.side}
\title{Internal function from mgcv}
\usage{
gam.side(sm, Xp, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{sm}{List of smooth terms.}

\item{Xp}{Parametric model matrix.}

\item{tol}{Numerical tolerance.}
}
\value{
A list of smooth terms, with identifiability constraints imposed.
}
\description{
This function is derived \code{mgcv:::gam.side}.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon Wood
}
\keyword{internal}
