% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm4-functions.R
\name{gamm4.setup}
\alias{gamm4.setup}
\title{Internal function for fitting GAMMs using lme4}
\usage{
gamm4.setup(formula, pterms, data = stop("No data supplied to gamm.setup"))
}
\arguments{
\item{pterms}{Parametric terms in the model.}

\item{data}{Model frame.}

\item{fixed}{A formula excluding lme4 style random effects but including
smooths.}
}
\value{
model
}
\description{
This function is derived from an internal function in the \code{gamm4} package.
}
\references{
\insertRef{woodStraightforwardIntermediateRank2013}{galamm}

\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\seealso{
\code{\link[=gamm4]{gamm4()}} and \code{\link[=gam.setup]{gam.setup()}}.
}
\author{
Simon N Wood, with some modifications by Oystein Sorensen.
}
\keyword{internal}
