% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature_rules.R
\name{generate_cubature_rules}
\alias{generate_cubature_rules}
\title{Generate cubature rules for integration}
\usage{
generate_cubature_rules(
  dimension,
  num_quadrature_points = NULL,
  type = "lu_darmofal_4.2"
)
}
\arguments{
\item{dimension}{Dimension of the integral.}

\item{num_quadrature_points}{Vector containing the number of
quadrature points per dimension. Only used when \code{type = "gauss-hermite"}.}

\item{type}{Type of rule.}
}
\value{
List of rules and weights.
}
\description{
Generate cubature rules for integration
}
