% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllamm.R
\name{gllamm}
\alias{gllamm}
\title{Fit the mixed model representation of a generalized
additive latent and mixed model}
\usage{
gllamm(
  formula,
  family = stats::gaussian(),
  data = list(),
  load_var = NULL,
  lambda = NULL,
  factor = NULL
)
}
\arguments{
\item{formula}{A formula for the fixed effects, including
smooth terms.}

\item{family}{A GLM family.}

\item{data}{A data frame containing the variables used in the
model.}

\item{load_var}{The variables which the factors load onto.}

\item{lambda}{A factor loading matrix, with row names corresponding
to \code{load_var}.}

\item{factor}{The latent variable, or factor.}

\item{random}{A formula for random effects.}
}
\value{
An object of S3 class \code{galamm}.
}
\description{
Experimental function which computes maximum
likelihood estimates of the mixed model
representation of a GALAMM. The function can also
be used for fitting GLLAMMs.
}
