% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hermite_quadrature}
\alias{hermite_quadrature}
\title{Gauss-Hermite quadrature rules}
\format{
A list with 100 elements. The \eqn{k}th element contains the order
\eqn{n} quadrature rules for the Hermite polynomials of order \eqn{k}. Each
list element is a dataframe containing the following elements:
\describe{
  \item{x}{Location of quadrature point}
  \item{w}{Quadrature weight}
}
}
\usage{
hermite_quadrature
}
\description{
List of Gauss-Hermite quadrature rules generated using the \code{gaussquad}
package \insertCite{novomestky2013}{galamm}, and then scaled by multiplying the
locations by \eqn{\sqrt{2}} and the weights by \eqn{1/\sqrt{pi}}, thus
making sure the standard normal distribution is being integrated over.
}
\references{
\insertAllCited{}
}
\keyword{internal}
