% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{interpret.gam0}
\alias{interpret.gam0}
\title{Internal function for interpreting GAM formulas}
\usage{
interpret.gam0(gf)
}
\arguments{
\item{gf}{A formula object.}
}
\value{
An object of class \code{split.gam.formula}, containing a
specification of the smooth terms and a specification of the
parametric components.
}
\description{
This formula is based on \code{mgcv:::interpret.gam0}. It takes a formula
and returns a model formula for the parametric and a list of
descriptors for the smooths.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon Wood, with modifications by Oystein Sorensen.
}
\keyword{internal}
