% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latent_covariates_long}
\alias{latent_covariates_long}
\title{Simulated Longitudinal Data with Latent and Observed Covariates Interaction}
\format{
\subsection{\code{latent_covariates_long} A data frame with 800 rows and 5}{

columns:
\describe{
\item{id}{Subject ID.}
\item{type}{Type of observation in the \code{y} variable.
If it equals \code{"measurement1"} or \code{"measurement2"} then the
observation is a measurement of the latent variable. If it equals
\code{"response"}, then the observation is the actual response.}
\item{x}{Explanatory variable.}
\item{y}{Observed response. Note, this includes both the actual response,
and the measurements of the latent variable, since mathematically they
are all treated as responses.}
\item{response}{Dummy variable indicating whether the given row is a
response or not.}
}
}
}
\usage{
latent_covariates_long
}
\description{
Simulated dataset for use in examples and testing with a latent covariate
interacting with an observed covariate. In this data, each response has been
measured six times for each subject.
}
\seealso{
Other datasets:
\code{\link{cognition}},
\code{\link{diet}},
\code{\link{epilep}},
\code{\link{hsced}},
\code{\link{latent_covariates}},
\code{\link{lifespan}},
\code{\link{mresp}},
\code{\link{mresp_hsced}}
}
\concept{datasets}
\keyword{datasets}
