% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Compute the marginal likelihood of a GLLAMM or GALAMM}
\usage{
marginal_likelihood(
  y,
  trials,
  X,
  Zt,
  Lambdat,
  beta,
  theta,
  theta_mapping,
  lambda,
  lambda_mapping_X,
  lambda_mapping_Zt,
  weights,
  weights_mapping,
  family,
  family_mapping,
  maxit_conditional_modes,
  hessian = FALSE,
  epsilon_u = 1e-10
)
}
\arguments{
\item{y}{A \code{numeric} vector of responses.}

\item{trials}{A \code{numeric} vector representing the number
of trials. Only used for binomial models, but should be set to
some arbitrary value otherwise.}

\item{X}{A dense \code{numeric} matrix.}

\item{Zt}{A sparse matrix of class \code{dgCMatrix}.}

\item{Lambdat}{A sparse matrix of class \code{dgCMatrix}.}

\item{beta}{A \code{numeric} vector of fixed effects.}

\item{theta}{A \code{numeric} vector of variance components,
parametrized identically to \code{lme4}.}

\item{theta_mapping}{An \code{integer} vector corresponding to
\code{Lind} used by \code{lme4}, but with base zero indexing.}

\item{lambda}{A \code{numeric} vector of factor loadings.}

\item{lambda_mapping_X}{An \code{integer} vector of mappings between
\code{X} and \code{lambda}, columnwise. Should be set to
\code{integer()} if not used. An entry \code{-1} indicates that the
corresponding value of \code{X} does not depend on \code{lambda},
as in the case where the first element of \code{lambda} is fixed to 1.}

\item{lambda_mapping_Zt}{An \code{integer} vector of mappings between
\code{Zt} and \code{lambda}, along the nonzero elements of \code{Zt}
as can be found by \code{Zt@x}. Should be set to
\code{integer()} if not used. An entry \code{-1} indicates that the
corresponding value of \code{X} does not depend on \code{lambda},
as in the case where the first element of \code{lambda} is fixed to 1.}

\item{weights}{Vector of weights.}

\item{weights_mapping}{Mapping}

\item{family}{A length one \code{character} denoting the family.}

\item{family_mapping}{Mapping}

\item{maxit_conditional_modes}{Maximum number of iterations for
conditional models. Can be 1 when \code{family = "gaussian"}.}

\item{hessian}{Boolean specifying whether to include the Hessian matrix
at the given parameters. Defaults to \code{FALSE}.}

\item{epsilon_u}{Tolerance in the inner iteration. Defaults to \code{1e-10}.}
}
\value{
A \code{list} with elements \code{logLik} and \code{gradient}.
}
\description{
This function computes the Laplace approximate marginal
likelihood.
}
\details{
For examples, see the vignette on maximum likelihood estimation.
}
