% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm_control.R
\name{new_galamm_control}
\alias{new_galamm_control}
\title{Constructor method for galamm_control objects}
\usage{
new_galamm_control(optim_control, maxit_conditional_modes)
}
\arguments{
\item{optim_control}{List passed on to \code{stats::optim}'s \code{control}
argument. If not otherwise specified, the following arguments are set to
non-default values: \code{fnscale = -1} and \code{lmm = 20}.}

\item{maxit_conditional_modes}{Maximum number of iterations in penalized
iteratively reweighted least squares algorithm. Ignored if \code{family =
  "gaussian"} for all observations, since then a single step gives the exact
answer.}
}
\value{
Object of class \code{galamm_control}, which typically will be
provided as an argument to \code{\link{galamm}}.
}
\description{
This function can be called for controling the optimization procedure used
when fitting GALAMMs using \code{\link{galamm}}.
}
\seealso{
\code{\link[=galamm_control]{galamm_control()}}
}
\keyword{internal}
