% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm_control.R
\name{optim_control.galamm_control}
\alias{optim_control.galamm_control}
\alias{optim_control}
\title{Extract optim control parameters}
\usage{
\method{optim_control}{galamm_control}(x)
}
\arguments{
\item{x}{Object of class galamm_control}
}
\value{
List
}
\description{
Extract optim control parameters
}
