% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{parametricPenalty}
\alias{parametricPenalty}
\title{Function from mgcv}
\usage{
parametricPenalty(pterms, assign, paraPen, sp0)
}
\arguments{
\item{pterms}{parametric terms}

\item{assign}{ss}

\item{paraPen}{ss}

\item{sp0}{ss}
}
\value{
value
}
\description{
This function is derived from an internal mgcv function, written by Simon Wood.
}
\author{
Simon Wood
}
