% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature_rules.R
\name{replicate_quadrature_points}
\alias{replicate_quadrature_points}
\title{Convencience function for replication quadrature points}
\usage{
replicate_quadrature_points(quadrature_rules, nrep)
}
\arguments{
\item{quadrature_rules}{Object returned by \code{\link{generate_cubature_rules}}.}

\item{nrep}{Number of repetitions.}
}
\value{
List of quadrature rules.
}
\description{
Convencience function for replication quadrature points
}
