% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooths.R
\name{s}
\alias{s}
\alias{t2}
\title{Set up smooth term with factor loading}
\usage{
s(..., load.var = NULL)

t2(..., load.var = NULL)
}
\arguments{
\item{...}{Arguments passed on to \code{mgcv::s} or \code{mgcv::t2}.}

\item{load.var}{Optional character argument specifying the loading variable.}
}
\value{
An object of class \code{xx.smooth.spec}, where \code{xx} is a basis
identifying code given by the \code{bs} argument of \code{s} or \code{t2}.
It differs from the smooth returned by \code{mgcv::s} or \code{mgcv::t2} in
that it has an additional attribute named \code{"load.var"} which specifies
any factor loading which this smooth term should be multiplied with in
order to produce the observed outcome.
}
\description{
Thes are very thin wrappers around \code{mgcv::s} and \code{mgcv::t2}. They
enable the specification of loading variables for smooth terms.
}
\references{
\insertRef{woodThinPlateRegression2003}{galamm}

\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
