% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.galamm}
\alias{sigma.galamm}
\title{Extract square root of dispersion parameter from galamm object}
\usage{
\method{sigma}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm}, returned from
\code{\link{galamm}}.}

\item{...}{Optional parameters passed on to other methods. Currently not
used.}
}
\value{
The square root of one or more dispersion parameters.
}
\description{
Extracts the square root of the dispersion parameter(s) from an object of
class \code{galamm}, returned from \code{\link{galamm}}. In the case of
conditionally Gaussian responses, this is the residual standard deviation.
When there are multiple dispersion parameters, e.g., with mixed response
type models, the square root of all of them are returned in a numeric vector.
}
\examples{
# Linear mixed model with heteroscedastic residuals
mod <- galamm(
  formula = y ~ x + (1 | id),
  dispformula = ~ (1 | item),
  data = hsced
)

# Extract residual standard deviation.
sigma(mod)

# The residual standard deviation applies to the base case. The variance
# function shown in the model output shows the estimated multiplier for
# various grouping levels:
summary(mod)

}
\seealso{
\code{\link[=galamm]{galamm()}}

Other details of model fit:
\code{\link{VarCorr}},
\code{\link[=appraise.galamm]{appraise.galamm()}},
\code{\link[=coef.galamm]{coef.galamm()}},
\code{\link[=confint.galamm]{confint.galamm()}},
\code{\link[=derivatives.galamm]{derivatives.galamm()}},
\code{\link[=deviance.galamm]{deviance.galamm()}},
\code{\link[=factor_loadings.galamm]{factor_loadings.galamm()}},
\code{\link[=family.galamm]{family.galamm()}},
\code{\link[=fitted.galamm]{fitted.galamm()}},
\code{\link{fixef}},
\code{\link[=formula.galamm]{formula.galamm()}},
\code{\link[=llikAIC]{llikAIC()}},
\code{\link[=logLik.galamm]{logLik.galamm()}},
\code{\link[=model.frame.galamm]{model.frame.galamm()}},
\code{\link[=nobs.galamm]{nobs.galamm()}},
\code{\link[=predict.galamm]{predict.galamm()}},
\code{\link[=print.VarCorr.galamm]{print.VarCorr.galamm()}},
\code{\link[=ranef.galamm]{ranef.galamm()}},
\code{\link[=residuals.galamm]{residuals.galamm()}},
\code{\link[=response]{response()}},
\code{\link[=vcov.galamm]{vcov.galamm()}}
}
\concept{details of model fit}
