% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_fixed}
\alias{update_fixed}
\title{Update fixed effect model matrix when loadings have changed}
\usage{
update_fixed(X, columns, lambda_update)
}
\arguments{
\item{X}{model matrix}

\item{columns}{integer vector of columns to update}

\item{lambda_update}{double vector of values with which to multiply old
values. One for each row of \code{X}.}
}
\value{
A matrix in which all columns specified in the
 \code{columns} argument have been divided by the values in
 \code{lambda_update}.
}
\description{
Update fixed effect model matrix when loadings have changed
}
