% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_random}
\alias{update_random}
\title{Update random effect model matrix when loadings have changed}
\usage{
update_random(Zt, lambda_update)
}
\arguments{
\item{Zt}{transposed model matrix in compressed sparse column format}

\item{lambda_update}{dense vector containing values with which the current
nonzero elements of \code{Zt} should be multiplied.}
}
\value{
A matrix having the same dimension and sparsity pattern as
\code{Zt}, but with its nonzero values updated.
}
\description{
Update random effect model matrix when loadings have changed
}
