% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{variable.summary}
\alias{variable.summary}
\title{Create variable summary}
\usage{
variable.summary(pf, dl, n)
}
\arguments{
\item{pf}{Formula for the parametric part of the model.}

\item{dl}{List containing all the explanatory variables in the model.}

\item{n}{Integer specifying the number of observations.}
}
\value{
A list containing summary statistics for each variable in the model.
}
\description{
This function is derived from \code{mgcv:::variable.summary}.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon Wood
}
\keyword{internal}
