% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{abbrDeparse}
\alias{abbrDeparse}
\title{Abbreviated deparse function taken from lme4}
\usage{
abbrDeparse(x, width = 60)
}
\arguments{
\item{x}{Name to deparse}

\item{width}{Field width}
}
\value{
String
}
\description{
Abbreviated deparse function taken from lme4
}
