% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{anova.galamm}
\alias{anova.galamm}
\title{Compare likelihoods of galamm objects}
\usage{
\method{anova}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{...}{Other fitted models of class \code{galamm}. Currently, if no
models are provided in this argument, no table will be returned.}
}
\value{
A table with model comparison metric.
}
\description{
Anova function for comparing different GALAMMs fitted on the same data.
}
\references{
\insertRef{batesFittingLinearMixedEffects2015}{galamm}
}
\seealso{
\code{\link[=summary.galamm]{summary.galamm()}} for the summary method and \code{\link[=anova]{anova()}} for the
generic function.
}
\author{
Some of the source code for this function is adapted from
\code{lme4:::anova.merMod}, with authors Douglas M. Bates, Martin Maechler,
Ben Bolker, and Steve Walker.
}
