% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{anova.galamm}
\alias{anova.galamm}
\title{Compare Likelihoods of galamm Objects}
\usage{
\method{anova}{galamm}(object, ...)
}
\arguments{
\item{object}{A model fitted by \code{\link{galamm}}}

\item{...}{Other optional fitted models of class \code{galamm}.}
}
\value{
A table
}
\description{
Compare Likelihoods of galamm Objects
}
\details{
Some of the source code for this function is adapted from
\code{lme4:::anova.merMod}, with authors Douglas M. Bates, Martin Maechler,
Ben Bolker, and Steve Walker.
}
