% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{augment.smX}
\alias{augment.smX}
\title{Internal function from mgcv}
\usage{
augment.smX(sm, nobs, np)
}
\arguments{
\item{sm}{smooth}

\item{nobs}{observations}

\item{np}{number}
}
\value{
value
}
\description{
Internal function from mgcv
}
\author{
Simon Wood
}
