% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.galamm}
\alias{coef.galamm}
\title{Extract galamm Coefficients}
\usage{
\method{coef}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{galamm}}}

\item{...}{Additional parameters}
}
\value{
An object
}
\description{
Extract galamm Coefficients
}
