% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.galamm}
\alias{coef.galamm}
\title{Extract galamm coefficients}
\usage{
\method{coef}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm}, from \code{\link{galamm}}.}

\item{...}{Additional parameters passed on to other methods. Currently not
used.}
}
\value{
A matrix with the requested coefficients.
}
\description{
Currently, this function only returns the fixed effects.
}
\seealso{
\code{\link[=fixef.galamm]{fixef.galamm()}} for fixed effects, \code{\link[=ranef.galamm]{ranef.galamm()}} for random
effects, and \code{\link[=coef]{coef()}} for the generic function.
}
