% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.galamm}
\alias{confint.galamm}
\title{Confidence intervals for model parameters}
\usage{
\method{confint}{galamm}(object, parm, level = 0.95, method = "Wald", ...)
}
\arguments{
\item{object}{Model object}

\item{parm}{Parameters for which to compute intervals. Use "theta" to get all
variance parameters, "beta" to get all fixed regression coefficients,
"lambda" to get all factor loadings, and "weights" to get all weights. Can
also be a vector of number.}

\item{level}{Confidence level.}

\item{method}{Currently only "Wald" is available.}

\item{...}{Other arguments.}
}
\value{
Confidence intervals.
}
\description{
Confidence intervals for model parameters
}
