% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Example dataset with three measurements of a single latent response variable.}
\format{
An object of class \code{data.frame} with 300 rows and 4 columns.
}
\usage{
dat1
}
\description{
The dataframe contains 300 observations of the following variables, for a total of
100 subjects:
\itemize{
\item{\code{id}: }{Subject ID, 1-100.}
\item{\code{item}: }{Item measured, 1-3.}
\item{\code{y}: }{Value of the measured response.}
\item{\code{x}: }{Predictor variable.}
}
The underlying smooth curve for the latent trait has been generated by
\code{mgcv::gamSim}.
}
\keyword{datasets}
