% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Example dataset with binomial response and two levels.}
\format{
A data frame with 300 rows and 4 variables:
\describe{
  \item{id}{Subject ID}
  \item{y}{Number of successes}
  \item{trials}{Number of trials. Always equal to 1.}
  \item{x}{Explanatory variable}
}
}
\usage{
dat1
}
\description{
Simulated dataset similar to the one used in Section 3.1 of
\insertCite{rabe-hesketh2005;textual}{galamm}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
