% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.galamm.R
\name{deviance.galamm}
\alias{deviance.galamm}
\title{Extract Deviance of galamm Object}
\usage{
\method{deviance}{galamm}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Other arguments}
}
\value{
Deviance
}
\description{
Extract Deviance of galamm Object
}
