% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_loadings.R
\name{factor_loadings.galamm}
\alias{factor_loadings.galamm}
\alias{factor_loadings}
\title{Extract Factor Loadings from galamm Object}
\usage{
\method{factor_loadings}{galamm}(object)
}
\arguments{
\item{object}{An object}
}
\value{
Vector
}
\description{
Extract Factor Loadings from galamm Object
}
\details{
This function has been named \code{factor_loadings} rather than just
\code{loadings} to avoid conflict with \code{stats::loadings}.
}
