% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.galamm.R
\name{fitted.galamm}
\alias{fitted.galamm}
\title{Extract model fitted values}
\usage{
\method{fitted}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
A numerical vector with fit values for each row in the input data.
}
\description{
Extract model fitted values
}
