% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.galamm.R
\name{fitted.galamm}
\alias{fitted.galamm}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{galamm}(object, ...)
}
\arguments{
\item{object}{galamm object}

\item{...}{other arguments}
}
\value{
Numerical vector
}
\description{
Extract Model Fitted Values
}
