% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef}
\alias{fixef}
\alias{fixef.galamm}
\title{Extract fixed effects from galamm objects}
\usage{
\method{fixef}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm}, returned from
\code{\link{galamm}}.}

\item{...}{Optional parameters passed on to other functions. Currently
not used.}
}
\value{
A matrix with the requested fixed effects.
}
\description{
Extract the fixed regression coefficients.
}
\examples{
# Poisson GLMM
count_mod <- galamm(formula = y ~ lbas * treat + lage + v4 + (1 | subj),
                    data = epilep, family = poisson)

# Extract fixed effects
fixef(count_mod)

}
\seealso{
\code{\link[=ranef.galamm]{ranef.galamm()}} for random effects, \code{\link[=coef.galamm]{coef.galamm()}} for
coefficients more generally, and \code{\link[=confint.galamm]{confint.galamm()}} for confidence intervals.
}
