% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.galamm}
\alias{formula.galamm}
\title{Extract formula from fitted galamm object}
\usage{
\method{formula}{galamm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{galamm} returned from \code{\link{galamm}}.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
The formula used to fit the model.
}
\description{
Extract formula from fitted galamm object
}
\examples{
# Mixed response model ------------------------------------------------------

# The mresp dataset contains a mix of binomial and Gaussian responses.

# We need to estimate a factor loading which scales the two response types.
loading_matrix <- matrix(c(1, NA), ncol = 1)

# Define mapping to families.
families <- c(gaussian, binomial)
family_mapping <- ifelse(mresp$itemgroup == "a", 1, 2)


# Fit the model
mod <- galamm(
  formula = y ~ x + (0 + level | id),
  data = mresp,
  family = families,
  family_mapping = family_mapping,
  factor = list("level"),
  load.var = "itemgroup",
  lambda = list(loading_matrix)
)

# Formula
formula(mod)

}
