% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm-package.R
\docType{package}
\name{galamm-package}
\alias{galamm-package}
\title{galamm: Estimation of Generalized Additive Latent and
Mixed Models.}
\description{
This package performs profile and maximum likelihood estmation of
generalized additive latent and mxied models.
}
