% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm-package.R
\docType{package}
\name{galamm-package}
\alias{galamm-package}
\title{galamm: Estimation of Generalized Additive Latent and
Mixed Models.}
\description{
This package performs maximum marginal likelihood estimation of
generalized additive latent and mixed models. To get started, take a look at
the vignettes. The main function is \code{\link{marginal_likelihood}}.
}
