% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fitting Generalized Additive Latent and Mixed Models}
\usage{
galamm(formula, data, family = gaussian, latent = NULL, lambda = NULL)
}
\arguments{
\item{formula}{a two-sided formula object using \code{lme4} style to specify
random effects.}

\item{data}{data frame containing the variables in \code{formula}.}

\item{family}{a GLM family, see \code{\link{glm}} and
\code{\link{family}}.}

\item{latent}{Formula for latent variables, of the form
\code{~(factor | load_var)}.}

\item{lambda}{Initial matrix for factor loadings.}
}
\value{
Object of class \code{galamm}.
}
\description{
Fitting Generalized Additive Latent and Mixed Models
}
\examples{
# empty example
}
