% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  weights = NULL,
  data,
  family = gaussian,
  family_mapping = rep(1L, nrow(data)),
  load.var = NULL,
  lambda = NULL,
  factor = NULL,
  start = NULL,
  control = galamm_control()
)
}
\arguments{
\item{formula}{A formula}

\item{weights}{An optional formula object specifying an expression for the
residual variance. Defaults to \code{NULL}, corresponding to homoscedastic
errors.}

\item{data}{A dataset}

\item{family}{A vector of families}

\item{family_mapping}{A vector mapping from the elements of "family" to rows
of "data". Defaults to \code{rep(1L, nrow(data))}, which means that all
observations are distributed according to the first element of "family".}

\item{load.var}{Variable the factors load onto}

\item{lambda}{Loading}

\item{factor}{list of factors}

\item{start}{A named list of starting values for parameters. Possible names
of list elements are "theta", "beta", and "lambda", all of which represent
numerical vectors.}

\item{control}{Control object. Result of calling
\code{\link{galamm_control}}.}
}
\value{
A model object
}
\description{
This function fits a generalized additive latent and mixed model (GALAMMs),
as described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm}.
The building blocks of these models are generalized additive mixed models
(GAMMs) \insertCite{woodGeneralizedAdditiveModels2017a;textual}{galamm}, of
which generalized linear mixed models
\insertCite{lairdRandomEffectsModelsLongitudinal1982,hendersonBestLinearUnbiased1975}{galamm}
are special cases. GALAMMs extend upon GAMMs by allowing factor structures,
as commonly used to model hypothesized latent traits underlying observed
measurements. In this sense, GALAMMs are an extension of generalized linear
latent and mixed models (GLLAMMs)
\insertCite{skrondalGeneralizedLatentVariable2004,rabe-heskethGeneralizedMultilevelStructural2004}{galamm}
which allows semiparametric estimation. The implemented algorithm used to
compute model estimates is described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm},
and is an extension of the algorithm used for fitting generalized linear
mixed models by the \code{lme4} package
\insertCite{batesFittingLinearMixedEffects2015}{galamm}. The syntax used to
define factor structures is based on that used by the \code{PLmixed} package,
which is detailed in
\insertCite{rockwoodEstimatingComplexMeasurement2019;textual}{galamm}.
}
\references{
\insertAllCited{}
}
