% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  data,
  family = gaussian,
  family_mapping = rep(1L, nrow(data)),
  load.var = NULL,
  lambda = NULL,
  factor = NULL
)
}
\arguments{
\item{formula}{A formula}

\item{data}{A dataset}

\item{family}{A vector of families}

\item{family_mapping}{A vector mapping from the elements of "family" to rows
of "data". Defaults to \code{rep(1L, nrow(data))}, which means that all
observations are distributed according to the first element of "family".}

\item{load.var}{Variable the factors load onto}

\item{lambda}{Loading}

\item{factor}{list of factors}
}
\value{
A model object
}
\description{
Fit a generalized additive latent and mixed model
}
