% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R, R/pkg.R
\docType{package}
\name{galamm}
\alias{galamm}
\title{Estimate a generalized additive latent and mixed model.}
\usage{
galamm(formula, data, family)
}
\arguments{
\item{formula}{The model formula, including smooth terms but not
random effects.}

\item{data}{Data frame or list.}

\item{family}{Family.}
}
\value{
An object of class galamm.
}
\description{
The package provides methods for estimating generalized additive mixed
models with mixed response types and latent variables.
}
