% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(formula, data, load.var = NULL, lambda = NULL, factor = NULL)
}
\arguments{
\item{formula}{A formula}

\item{data}{A dataset}

\item{load.var}{Variable the factors load onto}

\item{lambda}{Loading}

\item{factor}{list of factors}
}
\value{
A model object
}
\description{
Fit a generalized additive latent and mixed model
}
