% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  random = NULL,
  family = stats::gaussian(),
  data = list(),
  load.var = NULL,
  lambda = NULL,
  factor = NULL
)
}
\arguments{
\item{formula}{A formula for the fixed effects, including
smooth terms.}

\item{random}{A formula for random effects.}

\item{family}{A GLM family.}

\item{data}{A data frame containing the variables used in the
model.}

\item{load.var}{The variables which the factors load onto.}

\item{lambda}{A factor loading matrix, with row names corresponding
to \code{load.var}.}

\item{factor}{The latent variable, or factor.}
}
\value{
An object of S3 class \code{galamm}.
}
\description{
Fits a GALAMM
}
