% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  weights = NULL,
  data,
  family = gaussian,
  family_mapping = rep(1L, nrow(data)),
  load.var = NULL,
  lambda = NULL,
  factor = NULL,
  start = NULL,
  control = galamm_control()
)
}
\arguments{
\item{formula}{A formula specifying the model. Smooth terms are defined in
the style of the \code{mgcv} and \code{gamm4} packages, see
\insertCite{woodGeneralizedAdditiveModels2017a}{galamm} for an
introduction. Random effects are specified using \code{lme4} syntax,
which is described in detail in
\insertCite{batesFittingLinearMixedEffects2015}{galamm}. Factor loadings
will also be part of the model formula, and is based on the syntax of the
\code{PLmixed} package
\insertCite{rockwoodEstimatingComplexMeasurement2019}{galamm}.}

\item{weights}{An optional formula object specifying an expression for the
residual variance. Defaults to \code{NULL}, corresponding to homoscedastic
errors. The formula is defined in \code{lme4} style; see vignettes and
examples for details.}

\item{data}{A dataset containing all the variables specified by the model
formula, with the exception of factor loadings.}

\item{family}{A vector containing one or more model families. For each
element in \code{family} there should be a corresponding element in
\code{family_mapping} specifying which elements of the response are
conditionally distributed according to the given family. Currently family
can be one of \code{gaussian}, \code{binomial}, and \code{poisson}, and
only canonical link functions are supported.}

\item{family_mapping}{A vector mapping from the elements of \code{family} to
rows of \code{data}. Defaults to \code{rep(1L, nrow(data))}, which means
that all observations are distributed according to the first element of
\code{family}.}

\item{load.var}{Character specifying the name of the variable in \code{data}
identifying what the factors load onto. That is, each unique value of
\code{load.var} corresponds to a unique factor loading. Currently only a
single loading}

\item{lambda}{List of factor loading matrices. Numerical values indicate that
the given value is fixed, while \code{NA} means that the entry is a
parameter to be estimated.}

\item{factor}{List of character vectors identical to the factor loadings
specified in \code{formula}. For each list element, the \eqn{j}th entry in
the character vector corresponds to the \eqn{j}th column of the
corresponding matrix in \code{lambda}.}

\item{start}{A named list of starting values for parameters. Possible names
of list elements are \code{"theta"}, \code{"beta"}, \code{"lambda"}, and
\code{"weights"}, all of should be numerical vectors with starting values.
Default to \code{NULL}, which means that some relatively sensible defaults
are used.}

\item{control}{Control object for the optimization procedure of class
\code{galamm_control} resulting from calling \code{\link{galamm_control}}.}
}
\value{
A model object of class \code{galamm}, containing the following
elements:
\itemize{
\item \code{call} the matched call used when fitting the model.
\item \code{random_effects} a list containing the following two elements:
\itemize{
\item \code{b} random effects in original parametrization.
\item \code{u} random effects standardized to have identity covariance matrix.
}
\item \code{model} a list with various elements related to the model setup and fit:
\itemize{
\item \code{deviance} deviance of final model.
\item \code{deviance_residuals} deviance residuals of the final model.
\item \code{df} degrees of freedom of model.
\item \code{family} a list of one or more family objects, as specified in the \code{family}
arguments to \code{galamm}.
\item \code{fit} a numeric vector with fitted values.
\item \code{hessian} Hessian matrix of final model, i.e., the second derivative of the
log-likelihood with respect to all model parameters.
\item \code{lmod} Linear model object returned by \code{lme4::lFormula}, which
is used internally for setting up the models.
\item \code{loglik} Log-likelihood of final model.
\item \code{n} Number of observations.
\item \code{pearson_residual} Pearson residuals of final model.
\item \code{response} A numeric vector containing the response values used when
fitting the model.
\item \code{weights_object} Object with weights used in model fitting. Is \code{NULL}
when no weights were used.
}
\item \code{parameters} A list object with model parameters and related information:
\itemize{
\item \code{beta_inds} Integer vector specifying the indices of fixed regression
coefficients among the estimated model parameters.
\item \code{dispersion_parameter} One or more dispersion parameters of the final model.
\item \code{lambda_dummy} Dummy matrix of factor loadings, which shows the structure of the
loading matrix that was supplied in the \code{lambda} arguments.
\item \code{lambda_inds} Integer vector specifying the indices of factor loadings among
the estimated model parameters.
\item \code{parameter_estimates} Numeric vector of final parameter estimates.
\item \code{parameter_names} Names of all parameters estimates.
\item \code{theta_inds} Integer vector specifying the indices of variance components among
the estimated model parameters. Technically these are the entries of the
Cholesky decomposition of the covariance matrix.
\item \code{weights_inds} Integer vector specifying the indices of estimated weights
(used in heteroscedastic Gaussian models) among the estimated model
parameters.
}
\item \code{gam} List containing information about smooth terms in the model. If no
smooth terms are contained in the model, then it is \code{NULL}.
}
}
\description{
This function fits a generalized additive latent and mixed model (GALAMMs),
as described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm}.
The building blocks of these models are generalized additive mixed models
(GAMMs) \insertCite{woodGeneralizedAdditiveModels2017a}{galamm}, of which
generalized linear mixed models
\insertCite{breslowApproximateInferenceGeneralized1993,harvilleMaximumLikelihoodApproaches1977,hendersonBestLinearUnbiased1975,lairdRandomEffectsModelsLongitudinal1982}{galamm}
are special cases. GALAMMs extend upon GAMMs by allowing factor structures,
as commonly used to model hypothesized latent traits underlying observed
measurements. In this sense, GALAMMs are an extension of generalized linear
latent and mixed models (GLLAMMs)
\insertCite{skrondalGeneralizedLatentVariable2004,rabe-heskethGeneralizedMultilevelStructural2004}{galamm}
which allows semiparametric estimation. The implemented algorithm used to
compute model estimates is described in
\insertCite{sorensenLongitudinalModelingAgeDependent2023;textual}{galamm},
and is an extension of the algorithm used for fitting generalized linear
mixed models by the \code{lme4} package
\insertCite{batesFittingLinearMixedEffects2015}{galamm}. The syntax used to
define factor structures is based on that used by the \code{PLmixed} package,
which is detailed in
\insertCite{rockwoodEstimatingComplexMeasurement2019;textual}{galamm}.
}
\examples{
# Mixed response model ------------------------------------

# The mresp dataset contains a mix of binomial and Gaussian responses.

# We need to estimate a factor loading which scales the two response types.
loading_matrix <- matrix(c(1, NA), ncol = 1)

# Define mapping to families.
families <- c(gaussian, binomial)
family_mapping <- ifelse(mresp$itemgroup == "a", 1, 2)


# Fit the model
mod <- galamm(
  formula = y ~ x + (0 + level | id),
  data = mresp,
  family = families,
  family_mapping = family_mapping,
  factor = list("level"),
  load.var = "itemgroup",
  lambda = list(loading_matrix)
)

# Summary information
summary(mod)

# Generalized additive mixed model with factor structures -------------------

# The cognition dataset contains simulated measurements of three latent
# time-dependent processes, corresponding to individuals' abilities in
# cognitive domains. We focus here on the first domain, and take a single
# random timepoint per person:
dat <- subset(cognition, domain == 1)
dat <- split(dat, f = dat$id)
dat <- lapply(dat, function(x) x[x$timepoint \%in\% sample(x$timepoint, 1), ])
dat <- do.call(rbind, dat)
dat$item <- factor(dat$item)

# At each timepoint there are three items measuring ability in the cognitive
# domain. We fix the factor loading for the first measurement to one, and
# estimate the remaining two. This is specified in the loading matrix.
loading_matrix <- matrix(c(1, NA, NA), ncol = 1)

# We can now estimate the model.
mod <- galamm(
  formula = y ~ 0 + item + s(x, load.var = "loading") +
    (0 + loading | id),
  data = dat,
  load.var = "item",
  lambda = list(loading_matrix),
  factor = list("loading")
)

# We can plot the estimated smooth term
plot_smooth(mod, shade = TRUE)

}
\references{
\insertAllCited{}
}
\seealso{
Other modeling functions: 
\code{\link{galamm_control}()},
\code{\link{s}()},
\code{\link{t2}()}
}
\concept{modeling functions}
