% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R
\name{galamm}
\alias{galamm}
\title{Fit a generalized additive latent and mixed model}
\usage{
galamm(
  formula,
  weights = NULL,
  data,
  family = gaussian,
  family_mapping = rep(1L, nrow(data)),
  load.var = NULL,
  lambda = NULL,
  factor = NULL,
  start = NULL,
  control = galamm_control()
)
}
\arguments{
\item{formula}{A formula}

\item{weights}{An optional formula object specifying an expression for the
residual variance. Defaults to \code{NULL}, corresponding to homoscedastic
errors.}

\item{data}{A dataset}

\item{family}{A vector of families}

\item{family_mapping}{A vector mapping from the elements of "family" to rows
of "data". Defaults to \code{rep(1L, nrow(data))}, which means that all
observations are distributed according to the first element of "family".}

\item{load.var}{Variable the factors load onto}

\item{lambda}{Loading}

\item{factor}{list of factors}

\item{start}{A named list of starting values for parameters. Possible names
of list elements are "theta", "beta", and "lambda", all of which represent
numerical vectors.}

\item{control}{Control object. Result of calling
\code{\link{galamm_control}}.}
}
\value{
A model object
}
\description{
Fit a generalized additive latent and mixed model
}
