% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm.R, R/pkg.R
\docType{package}
\name{galamm}
\alias{galamm}
\title{Estimate a generalized additive latent and mixed model.}
\usage{
galamm(formula, data, family, nAGQ, maxit_newton = 10, maxit_moments = 10)
}
\arguments{
\item{formula}{The model formula, including smooth terms but not
random effects.}

\item{data}{Data frame or list.}

\item{family}{Family.}

\item{nAGQ}{Quadrature points. Either a vector of the same length as the
number of random effects, or a single value.}

\item{maxit_newton}{Newton iterations.}

\item{maxit_moments}{Moment iterations.}
}
\value{
An object of class galamm.
}
\description{
The package provides methods for estimating generalized additive mixed
models with mixed response types and latent variables.
}
\examples{
mod <- galamm(
  formula = y ~ x + (1 | id),
  data = dat1,
  family = binomial(),
  nAGQ = 4, maxit_newton = 2, maxit_moments = 3)
}
