% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galamm_control.R
\name{galamm_control}
\alias{galamm_control}
\title{Create new galamm_control object}
\usage{
galamm_control(trace = 3L, lmm = 20)
}
\arguments{
\item{trace}{Verbosity. Passed onto \code{\link{stats::optim}}}

\item{lmm}{Number of BFGS updates retained in the L-BFGS-B method.}
}
\value{
Object of class galamm_control
}
\description{
Create new galamm_control object
}
