% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{gam.setup}
\alias{gam.setup}
\title{Function for setting up GAM terms}
\usage{
gam.setup(
  formula,
  pterms,
  data = stop("No data supplied to gam.setup"),
  knots = NULL
)
}
\arguments{
\item{formula}{Formula including smooth terms, but not \code{lme4} style
random effect terms.}

\item{pterms}{Parametric terms.}

\item{data}{Model data.}

\item{knots}{Knots specification.}
}
\value{
A list containing all the data necessary to fit a GAMM.
}
\description{
This function is derived from \code{mgcv:::gam.setup}.
}
\references{
\insertRef{woodGeneralizedAdditiveModels2017a}{galamm}
}
\author{
Simon N Wood and Oystein Sorensen.
}
\keyword{internal}
