% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{gam.setup}
\alias{gam.setup}
\title{GAM setup function from mgcv}
\usage{
gam.setup(
  formula,
  pterms,
  data = stop("No data supplied to gam.setup"),
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  H = NULL,
  absorb.cons = TRUE,
  sparse.cons = 0,
  select = FALSE,
  idLinksBases = TRUE,
  scale.penalty = TRUE,
  paraPen = NULL,
  gamm.call = FALSE,
  drop.intercept = FALSE,
  diagonal.penalty = FALSE,
  apply.by = TRUE,
  list.call = FALSE,
  modCon = 0
)
}
\arguments{
\item{formula}{formula}

\item{pterms}{parametric terms}

\item{data}{data}

\item{knots}{knots}

\item{sp}{smoothing parameter}

\item{min.sp}{min}

\item{H}{hessian}

\item{absorb.cons}{logical}

\item{sparse.cons}{logical}

\item{select}{variable selection}

\item{idLinksBases}{ID}

\item{scale.penalty}{not documented yet}

\item{paraPen}{not documented yet}

\item{gamm.call}{not documented yet}

\item{drop.intercept}{not documented yet}

\item{diagonal.penalty}{not documented yet}

\item{apply.by}{not documented yet}

\item{list.call}{not documented yet}

\item{modCon}{not documented yet}
}
\value{
an object
}
\description{
This function is derived from mgcv's internal function with the same name.
}
\author{
Simon N Wood
}
