% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv.R
\name{gam.side}
\alias{gam.side}
\title{Internal function from mgcv}
\usage{
gam.side(sm, Xp, tol = .Machine$double.eps^0.5, with.pen = FALSE)
}
\arguments{
\item{sm}{list of smooth}

\item{Xp}{parametric model matrix}

\item{tol}{tolerance}

\item{with.pen}{logical}
}
\value{
a value
}
\description{
This function is derived from a function with the same name in mgcv, and is
written by Simon Wood.
}
\author{
Simon Wood
}
