% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm4-functions.R
\name{gamm4}
\alias{gamm4}
\title{Internal function for setting up model}
\usage{
gamm4(fixed, random = NULL, data = list())
}
\arguments{
\item{fixed}{formula excluding lme4 style random effects but including smooths}

\item{random}{random part}

\item{data}{data}
}
\value{
lmod object
}
\description{
Internal function for setting up model
}
\author{
Simon Wood and Fabian Scheipl
}
