% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm4.R
\name{gamm4}
\alias{gamm4}
\title{Internal function for setting up model}
\usage{
gamm4(
  formula,
  random = NULL,
  family = gaussian(),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action,
  knots = NULL,
  drop.unused.levels = TRUE,
  REML = FALSE,
  control = NULL,
  start = NULL,
  verbose = 0L,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{random}{random part}

\item{family}{model family}

\item{data}{data}

\item{weights}{weights}

\item{subset}{subset}

\item{na.action}{what to do}

\item{knots}{know}

\item{drop.unused.levels}{logical}

\item{REML}{for Gaussian model}

\item{control}{control}

\item{start}{starting point}

\item{verbose}{whether to blabla}

\item{...}{other arguments}
}
\value{
lmod object
}
\description{
Internal function for setting up model
}
\author{
Simon Wood and Fabian Scheipl
}
