% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm4-functions.R
\name{gamm4.setup}
\alias{gamm4.setup}
\title{Function to set up gamm4 model}
\usage{
gamm4.setup(
  formula,
  pterms,
  data = stop("No data supplied to gamm.setup"),
  knots = NULL
)
}
\arguments{
\item{formula}{formula}

\item{pterms}{parametric terms}

\item{data}{data}

\item{knots}{knots}
}
\value{
model
}
\description{
This function is derived from an internal function in the gamm4 package.
}
\author{
Simon N Wood
}
\keyword{internal}
