% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature_rules.R
\name{generate_cubature_rules}
\alias{generate_cubature_rules}
\title{Generate cubature rules for integration}
\usage{
generate_cubature_rules(
  dimension,
  num_quadrature_points = NULL,
  type = "lu_darmofal_4.2",
  weight_multiplier = pi^(-dimension/2),
  location_multiplier = sqrt(2)
)
}
\arguments{
\item{dimension}{Dimension of the integral.}

\item{num_quadrature_points}{Vector containing the number of
quadrature points per dimension. Only used when \code{type = "gauss-hermite"}.}

\item{type}{Type of rule.}

\item{weight_multiplier}{Number which the weights are multiplied by. Defaults
to yielding standard normal distribution.}

\item{location_multiplier}{Number which each location is multiplied with.}
}
\value{
List of rules and weights.
}
\description{
Generate cubature rules for integration
}
