% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsced}
\alias{hsced}
\title{Example Data with Heteroscedastic Residuals}
\format{
\subsection{\code{hsced} A data frame with 1200 rows and 5 columns:}{

\describe{
\item{id}{Subject ID.}
\item{age}{Timepoint.}
\item{item}{Item indicator.}
\item{x}{Explanatory variable}
\item{y}{Outcome.}
}
}
}
\usage{
hsced
}
\description{
Simulated dataset with residual standard deviation that varies between
items.
}
\references{
\insertAllCited{}
}
\seealso{
Other datasets: 
\code{\link{cognition}},
\code{\link{diet}},
\code{\link{epilep}},
\code{\link{latent_covariates_long}},
\code{\link{latent_covariates}},
\code{\link{mresp}}
}
\concept{datasets}
\keyword{datasets}
