% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsced}
\alias{hsced}
\title{Example Data with Heteroscedastic Residuals}
\format{
\subsection{\code{diet} A data frame with 1200 rows and 5 columns:}{

\describe{
\item{id}{Subject ID.}
\item{age}{Timepoint.}
\item{item}{Item indicator.}
\item{x}{Explanatory variable}
\item{y}{Outcome.}
}
}
}
\usage{
hsced
}
\description{
Simulated dataset with residual standard deviation that varies between
items.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
