% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{llikAIC.galamm}
\alias{llikAIC.galamm}
\title{Extract log likelihood, AIC, and related statistics from a GALAMM}
\usage{
\method{llikAIC}{galamm}(object)
}
\arguments{
\item{object}{Object of class \code{galamm} returned from
\code{\link{galamm}}.}
}
\value{
A list containing AIC, BIC, log likelihood, deviance and residual
degrees of freedom.
}
\description{
This function is assembles the values used by \code{\link{summary.galamm}}.
}
