% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.galamm.R
\name{logLik.galamm}
\alias{logLik.galamm}
\title{Extract Log-Likelihood of galamm Object}
\usage{
\method{logLik}{galamm}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Other arguments}
}
\value{
Object of class \code{logLik}
}
\description{
Extract Log-Likelihood of galamm Object
}
