% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.galamm.R
\name{logLik.galamm}
\alias{logLik.galamm}
\title{Extract Log-Likelihood of galamm Object}
\usage{
\method{logLik}{galamm}(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Other arguments}
}
\value{
Object of class \code{logLik}
}
\description{
Extract Log-Likelihood of galamm Object
}
\examples{
# Linear mixed model with heteroscedastic residuals
mod <- galamm(
  formula = y ~ x + (1 | id),
  weights = ~ (1 | item),
  data = hsced
)

# Extract log likelihood
logLik(mod)

}
\seealso{
\code{\link[=deviance.galamm]{deviance.galamm()}} for a function returning deviance and
\code{\link[=logLik]{logLik()}} for the generic function.
}
