% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_frame.R
\name{model.frame.galamm}
\alias{model.frame.galamm}
\title{Extract the model frame from a galamm object}
\usage{
\method{model.frame}{galamm}(formula, ...)
}
\arguments{
\item{formula}{An object of type \code{galamm} as described in \link{galammObject}.}

\item{...}{Other arguments. Currently not used.}
}
\value{
A data frame.
}
\description{
Extract the model frame from a galamm object
}
\seealso{
Other details of model fit: 
\code{\link{VarCorr}()},
\code{\link{coef.galamm}()},
\code{\link{confint.galamm}()},
\code{\link{derivatives.galamm}()},
\code{\link{deviance.galamm}()},
\code{\link{factor_loadings.galamm}()},
\code{\link{family.galamm}()},
\code{\link{fitted.galamm}()},
\code{\link{fixef}()},
\code{\link{formula.galamm}()},
\code{\link{llikAIC}()},
\code{\link{logLik.galamm}()},
\code{\link{nobs.galamm}()},
\code{\link{predict.galamm}()},
\code{\link{print.VarCorr.galamm}()},
\code{\link{ranef.galamm}()},
\code{\link{residuals.galamm}()},
\code{\link{response}()},
\code{\link{sigma.galamm}()},
\code{\link{vcov.galamm}()}
}
\concept{details of model fit}
