% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mresp}
\alias{mresp}
\title{Simulated Mixed Response Data}
\format{
\subsection{\code{mresp} A data frame with 2000 rows and 3 columns:}{

\describe{
\item{id}{Subject ID.}
\item{item}{Item indicator. Equals 1 for normally distributed response and 2 for binomially distributed response.}
\item{y}{Response.}
}
}
}
\usage{
mresp
}
\description{
Very basic mixed response dataset with one set of normally distributed
responses and one set of binomially distributed responses.
}
\keyword{datasets}
