% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{nobs.galamm}
\alias{nobs.galamm}
\title{Extract the Number of Observations from a galamm Fit}
\usage{
\method{nobs}{galamm}(object, ...)
}
\arguments{
\item{object}{galamm object}

\item{...}{other arguments}
}
\value{
A number
}
\description{
Extract the Number of Observations from a galamm Fit
}
\seealso{
Other details of model fit: 
\code{\link{VarCorr}()},
\code{\link{coef.galamm}()},
\code{\link{confint.galamm}()},
\code{\link{deviance.galamm}()},
\code{\link{factor_loadings.galamm}()},
\code{\link{family.galamm}()},
\code{\link{fitted.galamm}()},
\code{\link{fixef}()},
\code{\link{llikAIC}()},
\code{\link{logLik.galamm}()},
\code{\link{predict.galamm}()},
\code{\link{print.VarCorr.galamm}()},
\code{\link{ranef.galamm}()},
\code{\link{residuals.galamm}()},
\code{\link{sigma.galamm}()},
\code{\link{vcov.galamm}()}
}
\concept{details of model fit}
