% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.galamm.R
\name{nobs.galamm}
\alias{nobs.galamm}
\title{Extract the Number of Observations from a galamm Fit}
\usage{
\method{nobs}{galamm}(object, ...)
}
\arguments{
\item{object}{galamm object}

\item{...}{other arguments}
}
\value{
A number
}
\description{
Extract the Number of Observations from a galamm Fit
}
\keyword{internal}
