% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.galamm.R
\name{plot.galamm}
\alias{plot.galamm}
\title{Diagnostic Plots for galamm Objects}
\usage{
\method{plot}{galamm}(x, ...)
}
\arguments{
\item{x}{A galamm object}

\item{...}{Other arguments}
}
\value{
A plot is displayed
}
\description{
Diagnostic Plots for galamm Objects
}
