% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth.R
\name{plot_smooth.galamm}
\alias{plot_smooth.galamm}
\alias{plot_smooth}
\title{Plot smooth terms for galamm fits}
\usage{
\method{plot_smooth}{galamm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{...}{Other optional arguments, passed on to \code{mgcv::plot.gam}.}
}
\value{
A plot is displayed on the screen.
}
\description{
Plots smooth terms of a fitted \code{galamm} object. This function is a thin
wrapper around \code{mgcv::plot.gam}
\insertCite{woodGeneralizedAdditiveModels2017a}{galamm}.
}
\references{
\insertAllCited{}
}
