% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.galamm.R
\name{predict.galamm}
\alias{predict.galamm}
\title{Predictions from a model at new data values}
\usage{
\method{predict}{galamm}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{A fitted model object}

\item{newdata}{Data from for which to evaluate predictions. Defaults to
"NULL", which means that the predictions are evaluate at the data used to
fit the model.}

\item{type}{Type of prediction object to be returned.}

\item{...}{Other arguments.}
}
\value{
A number vector of predicted values.
}
\description{
Predictions from a model at new data values
}
