% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.galamm.R
\name{predict.galamm}
\alias{predict.galamm}
\title{Predictions from a model at new data values}
\usage{
\method{predict}{galamm}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{An object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{newdata}{Data from for which to evaluate predictions. Defaults to
"NULL", which means that the predictions are evaluate at the data used to
fit the model.}

\item{type}{Type of prediction object to be returned.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
A numeric vector of predicted values.
}
\description{
Currently, predictions are given with all random effects set to zero. We will
add the opportunity to specify the level of prediction in the future. This is
particularly important for models containing smooth terms, for which the
predictions currently don't make much sense.
}
\examples{
# Poisson GLMM
count_mod <- galamm(formula = y ~ lbas * treat + lage + v4 + (1 | subj),
                    data = epilep, family = poisson)

# Plot response versus link:
plot(
     predict(count_mod, type = "link"),
     predict(count_mod, type = "response")
     )

}
\seealso{
\code{\link[=fitted.galamm]{fitted.galamm()}} for model fits, \code{\link[=residuals.galamm]{residuals.galamm()}} for residuals,
and \code{\link[=predict]{predict()}} for the generic function.
}
