% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.galamm.R
\name{predict.galamm}
\alias{predict.galamm}
\title{Predictions from a model at new data values}
\usage{
\method{predict}{galamm}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{An object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{newdata}{Data from for which to evaluate predictions. Defaults to
"NULL", which means that the predictions are evaluate at the data used to
fit the model.}

\item{type}{Type of prediction object to be returned.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
A numeric vector of predicted values.
}
\description{
Predictions are given at the population level, i.e., with random effects set
to zero. For mixed response models, only predictions on the scale of the
linear predictors is supported.
}
\examples{
# Poisson GLMM
count_mod <- galamm(
  formula = y ~ lbas * treat + lage + v4 + (1 | subj),
  data = epilep, family = poisson
)

# Plot response versus link:
plot(
  predict(count_mod, type = "link"),
  predict(count_mod, type = "response")
)

}
\seealso{
\code{\link[=fitted.galamm]{fitted.galamm()}} for model fits, \code{\link[=residuals.galamm]{residuals.galamm()}} for
residuals, and \code{\link[=predict]{predict()}} for the generic function.

Other details of model fit: 
\code{\link{VarCorr}()},
\code{\link{coef.galamm}()},
\code{\link{confint.galamm}()},
\code{\link{deviance.galamm}()},
\code{\link{factor_loadings.galamm}()},
\code{\link{family.galamm}()},
\code{\link{fitted.galamm}()},
\code{\link{fixef}()},
\code{\link{llikAIC}()},
\code{\link{logLik.galamm}()},
\code{\link{print.VarCorr.galamm}()},
\code{\link{ranef.galamm}()},
\code{\link{residuals.galamm}()},
\code{\link{sigma.galamm}()},
\code{\link{vcov.galamm}()}
}
\concept{details of model fit}
