% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{print.summary.galamm}
\alias{print.summary.galamm}
\title{Print method for GALAMM fits}
\usage{
\method{print}{summary.galamm}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class "summary.galamm".}

\item{digits}{Number of digits.}

\item{...}{Further arguments.}
}
\value{
Summary printed to screen. Invisible returns the argument \code{x}.
}
\description{
Print method for GALAMM fits
}
