% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{print.summary.galamm}
\alias{print.summary.galamm}
\title{Print method for summary outputs}
\usage{
\method{print}{summary.galamm}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{\link{galamm}}}

\item{digits}{minimal number of significant digits.}

\item{...}{Additional arguments.}
}
\description{
Print the summary output of a fitted \code{\link{galamm}} object.
}
