% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{print.VarCorr.galamm}
\alias{print.VarCorr.galamm}
\title{Print method for variance-covariance objects}
\usage{
\method{print}{VarCorr.galamm}(
  x,
  digits = max(3, getOption("digits") - 2),
  comp = c("Std.Dev.", "Variance"),
  corr = any(comp == "Std.Dev."),
  ...
)
}
\arguments{
\item{x}{An object of class \code{VarCorr.galamm}, returned from
\code{\link{VarCorr.galamm}}.}

\item{digits}{Optional arguments specifying number of digits to use when
printing.}

\item{comp}{Character vector of length 1 or 2 specifying which variance
components to print.}

\item{corr}{Boolean indicating whether covariances or correlations should be
printed.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
The variance-covariance information is printed to the console and the
argument \code{x} is silently returned.
}
\description{
Print method for variance-covariance objects
}
\references{
\insertRef{batesFittingLinearMixedEffects2015}{galamm}
}
\seealso{
\code{\link[=VarCorr.galamm]{VarCorr.galamm()}} for the function creating the variance-covariance
objects.
}
\author{
This function is derived from \code{lme4:::print.VarCorr.merMod}
written by Douglas M. Bates, Martin Maechler, Ben Bolker, and Steve Walker.
}
