% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.galamm}
\alias{ranef.galamm}
\alias{ranef}
\title{Extract Random Effects from galamm Object}
\usage{
\method{ranef}{galamm}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{
Random effects
}
\description{
Extract Random Effects from galamm Object
}
\author{
This function is derived from \code{lme4::ranef.merMod}, written by
Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker.
}
