% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.galamm.R
\name{residuals.galamm}
\alias{residuals.galamm}
\title{Residuals of galamm Objects}
\usage{
\method{residuals}{galamm}(object, type = c("pearson", "deviance"), ...)
}
\arguments{
\item{object}{A galamm object}

\item{type}{Type}

\item{...}{Other arguments}
}
\value{
Vector
}
\description{
Residuals of galamm Objects
}
