% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{summary.galamm}
\alias{summary.galamm}
\title{Summary method for GALAMM fits}
\usage{
\method{summary}{galamm}(object, ...)
}
\arguments{
\item{object}{A fitted \code{\link{galamm}} object.}

\item{...}{Other arguments.}
}
\value{
Summary information.
}
\description{
Summary information for a fitted \code{\link{galamm}} object.
}
\examples{
# Example dataset with three measurements of a latent response
# Define loading matrix
load.mat <- matrix(c(1, NA, NA), ncol = 1)
dimnames(load.mat) <- list(c("item1", "item2", "item3"), NULL)
\dontrun{
mod <- galamm(
  formula = y ~ s(x, by = weight),
  random = ~(1|id),
  data = dat1,
  load_var = "item",
  lambda = load.mat,
  factor = "weight",
  lambda_init = c(2, .4)
  )
x <- summary(mod)
  }

}
\author{
Much of the code is based on the \code{summary} function of the
\code{PLmixed} package, written by Minjeong Jeon and Nicholas Rockwood.
}
