% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{summary.galamm}
\alias{summary.galamm}
\title{Summarizing GALAMM Fits}
\usage{
\method{summary}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class "galamm", usually the result of calling
\code{\link{galamm}}.}

\item{...}{Further arguments.}
}
\value{
A list of summary statistics of the fitted model.
}
\description{
Summary method for class "galamm".
}
