% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.galamm.R
\name{summary.galamm}
\alias{summary.galamm}
\title{Summarizing GALAMM fits}
\usage{
\method{summary}{galamm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{galamm} returned from
\code{\link{galamm}}.}

\item{...}{Further arguments passed on to other methods. Currently not used.}
}
\value{
A list of summary statistics of the fitted model, of class
\code{summary.galamm}.
}
\description{
Summary method for class "galamm".
}
\examples{
# Linear mixed model with heteroscedastic residuals
mod <- galamm(
  formula = y ~ x + (1 | id),
  weights = ~ (1 | item),
  data = hsced
)

summary(mod)

}
\seealso{
\code{\link[=print.summary.galamm]{print.summary.galamm()}} for the print method and \code{\link[=summary]{summary()}} for
the generic.

Other {summary functions}: 
\code{\link{anova.galamm}()},
\code{\link{plot.galamm}()},
\code{\link{plot_smooth.galamm}()},
\code{\link{print.summary.galamm}()}
}
\author{
Some of the code for producing summary information has been derived
from the summary methods of \code{mgcv} (author: Simon Wood) and
\code{lme4} \insertCite{batesFittingLinearMixedEffects2015}{galamm}
(authors: Douglas M. Bates, Martin Maechler, Ben Bolker, and Steve Walker).
}
\concept{{summary functions}}
