% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-functions.R
\name{variable.summary}
\alias{variable.summary}
\title{Internal function from mgcv}
\usage{
variable.summary(pf, dl, n)
}
\arguments{
\item{pf}{pd}

\item{dl}{dl}

\item{n}{n}
}
\value{
summary
}
\description{
This function is derived from an internal mgcv function, written by Simon Wood.
}
\author{
Simon Wood
}
