% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.galamm}
\alias{vcov.galamm}
\title{Calculate Variance-Covariance Matrix for galamm Model Object}
\usage{
\method{vcov}{galamm}(object, parm = "beta", ...)
}
\arguments{
\item{object}{Fitted model}

\item{parm}{The parameters for which the variance-covariance matrix should
be calculated. Character vector with one or more of the elements "theta",
"beta", "lambda", and "weights". Can also be an integer vector.}

\item{...}{Further arguments}
}
\value{
Variance-covariance matrix
}
\description{
Calculate Variance-Covariance Matrix for galamm Model Object
}
